Fix compile warnings in MSVC

+++ dynet/weight-decay.h
@@ -23,7 +23,7 @@ struct L2WeightDecay {
     if (num_updates == 0) return;
     if (num_updates == 1)
       weight_decay -= weight_decay * lambda;
-    else weight_decay = weight_decay * std::pow(1-lambda, num_updates);
+	else weight_decay = (float)(weight_decay * std::pow(1 - lambda, num_updates));
   }
   float current_weight_decay() const { return weight_decay; }
   bool parameters_need_rescaled() const {
